new structure admin panel tests + Machinery tests
